requirements.txt
qsharp
qsharp-widgets
ipykernel
azure-quantum
pulser
pulser-pasqal
matplotlib
pylatexenc
pyperclip
